# [python] Use non-empty domain to significantly improve performance on to-CSR