**Issue and/or context:** Found during performance analysis.

**Changes:** Now that we can have "room for growth" large shape-as-capacity, performance of the `to_csr` can be improved by orders of magnitude when we utiilize bounding-box data.

**Notes for Reviewer:**

Analysis script: https://gist.github.com/johnkerl/6d662b928938379209c53a9f793d0ecc

Timings before this PR:

```
$ analysis.py tiledb://johnkerl-tiledb/tabula-sapiens-immune-registered
URI tiledb://johnkerl-tiledb/tabula-sapiens-immune-registered
---------------------------------------------------------------- get obs/var lengths
 7.477 SECONDS MISC. nobs: 1000 nvar: 30000

---------------------------------------------------------------- query X blockwise CSR
51.985 seconds X block-iter page shape: (1000, 30000)
58.863 SECONDS X BLOCK-ITER DOUBLE-SELECT TOTAL

---------------------------------------------------------------- query X tiledb row-sel scipy col-sel
36.152 seconds X read shape: (2147483646, 2147483646)
 0.014 seconds X to_scipy shape: (2147483646, 2147483646)
 6.870 seconds X to_csr shape: (2147483646, 2147483646) nnz: 2184147
14.412 seconds X var subset shape: (2147483646, 30000) nnz: 1128774
57.448 SECONDS X TILEDB ROW-SEL SCIPY COL-SEL TOTAL

---------------------------------------------------------------- query X tiledb row/col-sel
37.402 seconds X double-select read
 0.004 seconds X double-select to_scipy shape: (2147483646, 2147483646) nnz: 1128774
 6.819 seconds X double-select to_csr nnz: 1128774
44.225 SECONDS X DOUBLE-SELECT TOTAL
```

Timings with this PR:

```
$ analysis.py tiledb://johnkerl-tiledb/tabula-sapiens-immune-registered
URI tiledb://johnkerl-tiledb/tabula-sapiens-immune-registered
---------------------------------------------------------------- get obs/var lengths
 7.414 SECONDS MISC. nobs: 1000 nvar: 30000

---------------------------------------------------------------- query X blockwise CSR
29.983 seconds X block-iter page shape: (1000, 30000)
38.550 SECONDS X BLOCK-ITER DOUBLE-SELECT TOTAL

---------------------------------------------------------------- query X tiledb row-sel scipy col-sel
22.756 seconds X read shape: (264824, 58559)
 0.015 seconds X to_scipy shape: (264824, 58559)
 0.019 seconds X to_csr shape: (264824, 58559) nnz: 2146409
 0.022 seconds X var subset shape: (264824, 30000) nnz: 1100047
22.812 SECONDS X TILEDB ROW-SEL SCIPY COL-SEL TOTAL

---------------------------------------------------------------- query X tiledb row/col-sel
21.030 seconds X double-select read
 0.004 seconds X double-select to_scipy shape: (264824, 58559) nnz: 1100047
 0.010 seconds X double-select to_csr nnz: 1100047
21.044 SECONDS X DOUBLE-SELECT TOTAL
```