@johnkerl 2024-01-26
Update:

* There have been discussions about dim-expansion (outside the scope of this PR)
* The following can be used to speed up to-CSR (without need for a PR):

```diff
exp = soma.open(uri)
meas = "RNA"
X_layer = exp.ms[meas].X["data"]
< xcoo = X_layer.read().coos().concat()
> nobs = exp.obs.count
> nvar = exp.ms[meas].var.count
> xcoo = X_layer.read().coos(shape=(nobs,nvar)).concat()
```